commitcode

+++ Models/DTO/Booking/ServiceBookingDetailDto.cs
@@ -5,6 +5,6 @@ namespace GraduationThesis_CarServices.Models.DTO.Booking
     {
         public string ServiceName { get; set; }
         public string ServiceImage { get; set; }
-        public float ServicePrice { get; set; }
+        public double ServicePrice { get; set; }
     }
 }
