+++ Models/DTO/ServiceDetail/ServiceDetailCreateRequestDto.cs
@@ -4,7 +4,7 @@ namespace GraduationThesis_CarServices.Models.DTO.ServiceDetail
 {
     public class ServiceDetailCreateRequestDto
     {
-        public float ServicePrice { get; set; }
+        public double ServicePrice { get; set; }
         public int MinNumberOfCarLot { get; set; }
         public int MaxNumberOfCarLot { get; set; }
         public int ServiceId { get; set; }
